change purchase_date to bigint

+++ src/db/scripts/create_ledger_table.sql
@@ -3,6 +3,6 @@ CREATE TABLE IF NOT EXISTS `ledger` (
   user varchar(255) NOT NULL,
   item varchar(255) NOT NULL,
   cost INT NOT NULL,
-  purchase_date INT NOT NULL,
+  purchase_date BIGINT NOT NULL,
   cleared TINYINT(1) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
